% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_on_maps.R
\name{add_cmarker}
\alias{add_cmarker}
\title{Draw a Marker}
\usage{
add_cmarker(mp, cx, cv, coords = NULL)
}
\arguments{
\item{mp}{a leaflet object}

\item{cx}{the properties of the centroid (see the \code{centroids} table)}

\item{coords}{a vector of coordinates values for the centroids}
}
\value{
a leaflet object
}
\description{
Draw a Marker for a centroid
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
