% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_on_maps.R
\name{add_cmarkers}
\alias{add_cmarkers}
\title{Draw all Markers}
\usage{
add_cmarkers(mp, cv)
}
\arguments{
\item{mp}{a leaflet object}

\item{cv}{a vector of coordinates values for the centroids, conveniently named}
}
\value{
a leaflet object
}
\description{
Draw a Marker for all centroids
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
