% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_on_maps.R
\name{add_poly_msoa}
\alias{add_poly_msoa}
\title{Draw MSOA polygon}
\usage{
add_poly_msoa(mp, yb, mn, mdn, mv, mc, mw, hc = "white", hw = 6)
}
\arguments{
\item{mp}{a leaflet object}

\item{yb}{the boundaries of the \code{MSOA} in \code{sf} format}

\item{mn}{the name of the \code{MSOA}}

\item{mdn}{the \emph{display} name for the Area}

\item{mv}{the values of the population for the cells included in the \code{MSOA}}

\item{mc}{the colour for the polygon line}

\item{mw}{the weight for the polygon line}

\item{hc}{the colour for the line when the polygon is highlighted}

\item{hw}{the weight for the line when the polygon is highlighted}
}
\value{
a leaflet object
}
\description{
Add an MSOA polygon on a \emph{leaflet} map as a polylines object
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
