% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_on_maps.R
\name{add_ppop}
\alias{add_ppop}
\title{Draw population points}
\usage{
add_ppop(mp, pb, pr, pn, rv, dn)
}
\arguments{
\item{mp}{a leaflet object}

\item{pb}{the points to be displayed in \code{sf} format. There must also be included a column called \code{pop} with the values to build a palette.}

\item{pr}{the radius of the points}

\item{pn}{the name of the \emph{ColorBrewer} palette (see \code{palettes.lst})}

\item{rv}{if \code{TRUE} the colours in the palette must be reversed}

\item{dn}{the name of the population segment}
}
\value{
a leaflet object
}
\description{
Add a layer of points representing a high density population grid related to a MSOA
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
