% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{dd_colour}
\alias{dd_colour}
\title{dd_colour}
\usage{
dd_colour(id, lbl, slz, full = TRUE)
}
\arguments{
\item{id}{id for the  control}

\item{lbl}{label for the control}

\item{slz}{colour scheme at start}

\item{full}{if TRUE returns the full object, otherwise the minimal compact}
}
\value{
a colorPickr control
}
\description{
A tailored \code{colorPickr} control from the \code{shinyWidgets} package
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
