% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down_zip.R
\name{down_zip}
\alias{down_zip}
\title{Download zip files}
\usage{
down_zip(x)
}
\arguments{
\item{x}{the folder where the unzipped csv files will be saved}
}
\value{
none
}
\description{
Download and unzip HDX/Meta High Resolution Population Density Files
}
\references{
For further details, see the
\href{https://data.humdata.org/dataset/united-kingdom-high-resolution-population-density-maps-demographic-estimates}{HDX dedicated webpage}
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
