% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{end_spinmap}
\alias{end_spinmap}
\title{end_spinmap}
\usage{
end_spinmap(mp, y = centers.uk$EW)
}
\arguments{
\item{mp}{a leaflet object}

\item{y}{the reference centroid (longitude and latitude, in this order)}
}
\value{
a leaflet object
}
\description{
end the \emph{spin plugin} in a \code{leaflet} map embedded in \code{Shiny} apps
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
