% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_on_maps.R
\name{grp_name}
\alias{grp_name}
\title{Name a leaflet marker}
\usage{
grp_name(x)
}
\arguments{
\item{x}{a row of the \code{centroids} table or all the table}
}
\value{
an HTML string to use as group name for the centroid markers
}
\description{
Create a name for each centroid marker or all at once
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
