% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_on_maps.R
\name{leaf_pal}
\alias{leaf_pal}
\title{leaflet palette}
\usage{
leaf_pal(pn, pv, rv = FALSE)
}
\arguments{
\item{pn}{the name of the \emph{ColorBrewer} palette (see \code{palettes.lst})}

\item{pv}{the vector values on which calculate the palette}

\item{rv}{if \code{TRUE} the colours in the palette must be reversed}
}
\value{
a leaflet palette object
}
\description{
Build a leaflet palette
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
