% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\docType{data}
\name{leafletspinPlugin}
\alias{leafletspinPlugin}
\title{leafletspinPlugin}
\format{
An object of class \code{html_dependency} of length 10.
}
\usage{
leafletspinPlugin
}
\description{
Dependency (2) needed for the spin plugin in leaflet maps during a proxy in shiny apps
}
\references{
\url{https://github.com/makinacorpus/Leaflet.Spin}
}
\keyword{datasets}
