% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_on_maps.R
\name{mod_map}
\alias{mod_map}
\title{Modify a leaflet map}
\usage{
mod_map()
}
\value{
a leaflet object
}
\description{
Start the process to modify the leaflet map in the shiny app, using the trick to trigger the \emph{spin} plugin
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
