% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{msoa.lst}
\alias{msoa.lst}
\title{msoa.lst}
\format{
An object of class \code{list} of length 7264.
}
\usage{
msoa.lst
}
\description{
List of Medium Super Output Areas (MSOA) names and ONS codes in England and Wales for use in a Shiny \emph{select} control.
The name of each Zone is completed with the name in brackets of the correspondent \code{LTLA}
}
\details{
For more details, consult the \href{https://geoportal.statistics.gov.uk/search?sort=-modified&tags=msoa}{ONS Geoportal} and
the \href{https://houseofcommonslibrary.github.io/msoanames/]}{House of Commons Library MSOA Names}
}
\keyword{datasets}
