% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{on_render_spin}
\alias{on_render_spin}
\title{on_render_spin}
\usage{
on_render_spin(mp, x = "")
}
\arguments{
\item{mp}{a leaflet object}

\item{x}{additional functionality to be executed after having stopped the spin plugin}
}
\value{
none
}
\description{
Helper JS function to (de)activate the spin plugin in leaflet maps during a proxy in shiny apps
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
