% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{palettes.lst}
\alias{palettes.lst}
\title{palettes.lst}
\format{
An object of class \code{list} of length 3.
}
\usage{
palettes.lst
}
\value{
a list
}
\description{
List of \emph{ColorBrewer} palettes grouped by type of visualization scale:
\itemize{
\item \strong{sequential}:  ordinal data where \emph{low} is less/more important and \emph{high} is more/less important
\item \strong{diverging}:   ordinal data where both low and high are equally but opposite important (deviation from some reference \emph{average} point)
\item \strong{qualitative}: categorical/nominal data where there is no apparent or logical order
}
}
\references{
\url{https://colorbrewer2.org/}
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
\keyword{datasets}
