% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_csv.R
\name{process_csv}
\alias{process_csv}
\title{Process CSV Files}
\usage{
process_csv(in_path, out_path = in_path)
}
\arguments{
\item{in_path}{the folder that stores the unzipped csv files}

\item{out_path}{the folder where the binary \code{fst} files will be saved}
}
\value{
none
}
\description{
Download HDX/Meta High Resolution Population Density Files
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
