% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{regPlugin}
\alias{regPlugin}
\title{regPlugin}
\usage{
regPlugin(mp, plugin)
}
\arguments{
\item{mp}{a leaflet object}

\item{plugin}{the definition of a \code{leaflet} plugin}
}
\value{
a leaflet object
}
\description{
Helper function to register plugins in \code{leaflet} maps
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
