% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RgridpopUK.R
\docType{package}
\name{RgridpopUK}
\alias{RgridpopUK}
\title{RgridpopUK}
\description{
A few functionalities on how to download and prepare data about Population in England and Wales at micro grid levels of 30 meters,
as published by the \href{https://data.humdata.org/}{Humanitarian Data Exchange}
on behalf of \href{https://data.humdata.org/organization/74ad0574-923d-430b-8d52-ad80256c4461}{Data for Good at Meta (previously Facebook)},
either as a total or segmented by the following characteristics:
\itemize{
\item Women
\item Men
\item Children (ages 0-5)
\item Youth (ages 15-24)
\item Elderly (ages 60+)
\item Women of reproductive age (ages 15-49)
}
}
\details{
The package also contains a processed and \emph{simplified} version, converted in \code{sf} format suitable for $R$ and Shiny,
of the Census 2021 boundaries for Middle-Layer Super Output Areas \code{MSOA}
(see the \href{https://www.ons.gov.uk/methodology/geography/ukgeographies/censusgeographies/census2021geographies}{ONS website}
for more information about the Census hierarchy geographic structure and its
\href{https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=name&tags=all(BDY_MSOA\%2CDEC_2021)}{geoportal}
for the original geographic files),
with their names aptly substituted as provided by the \href{https://houseofcommonslibrary.github.io/msoanames/}{House of Commons Library}.

A Shiny app is also included ready to run as soon as the data have been processed.
\subsection{Attributions}{
\itemize{
\item Facebook Connectivity Lab and Center for International Earth Science Information Network - CIESIN - Columbia University. 2016. High Resolution Settlement Layer (HRSL). Source imagery for HRSL © 2016 DigitalGlobe. Accessed 14 Dec 2022."
\item Contains OS data © Crown copyright and database rights \link{2022}
\item Contains National Statistics data © Crown copyright and database rights \link{2022}
\item Source: Office for National Statistics licensed under the \href{http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/}{Open Government Licence v.3.0}
\item Contains Parliamentary information licensed under the \href{https://www.parliament.uk/site-information/copyright/open-parliament-licence/}{Open Parliament Licence v3.0}
}
}
}
