% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\docType{data}
\name{spinPlugin}
\alias{spinPlugin}
\title{spinPlugin}
\format{
An object of class \code{html_dependency} of length 10.
}
\usage{
spinPlugin
}
\description{
Dependency (1) needed for the spin plugin in leaflet maps during a proxy in shiny apps
}
\references{
\url{https://spin.js.org/}
}
\keyword{datasets}
