% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{spop.lst}
\alias{spop.lst}
\title{spop.lst}
\format{
An object of class \code{character} of length 7.
}
\usage{
spop.lst
}
\description{
List of the various population segments available in the datasets
}
\keyword{datasets}
