% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maptiles.R
\docType{data}
\name{tiles.lst}
\alias{tiles.lst}
\title{tiles.lst}
\format{
An object of class \code{list} of length 21.
}
\usage{
tiles.lst
}
\value{
a list
}
\description{
List of background tiles for \code{leaflet} maps
}
\references{
\url{https://leaflet-extras.github.io/leaflet-providers/preview/}.

To work correctly, needs to be paired with \code{add_maptile}.
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
\keyword{datasets}
