% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroids.R
\name{wcentroids}
\alias{wcentroids}
\title{wcentroids}
\usage{
wcentroids(
  y,
  crs,
  coords = c("x_lon", "y_lat"),
  from_wgs = FALSE,
  cn = "pop",
  id = NULL,
  fexp = 1,
  prfx = "w"
)
}
\arguments{
\item{y}{a data.table}

\item{crs}{the projection CRS of the data in input, if any (see \code{from_wgs})}

\item{coords}{the name of the columns with the coordinates (first longitude, then latitude)}

\item{from_wgs}{if \code{TRUE} accept the input as from geographic coordinates}

\item{cn}{the name of the column with the values to be weighted}

\item{id}{an optional column name to use for grouping}

\item{fexp}{a value to add as exponent (fexp==1, mean; fexp=0.5, square root; fexp=2, quad mean; and so on)}

\item{prfx}{a prefix to add to \code{coords} in output}
}
\value{
a data.table
}
\description{
Calculates a series of \emph{weighted} centroids with an output in CRS WGS84
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
