% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst.R
\name{write_fst_idx}
\alias{write_fst_idx}
\title{Write a dataset in fst format with an index over one of its columns}
\usage{
write_fst_idx(dts, fname, cname, out_path)
}
\arguments{
\item{dts}{The dataset to be saved. If \code{NA}, both \code{dname} and \code{tname} must be provided.}

\item{fname}{The name of the file in output or the table to be read if \code{dname} is not \code{NA}}

\item{cname}{The name of the column on which to order the dataset and create the index}

\item{out_path}{The folder where to save the file.}
}
\value{
None
}
\description{
Write a dataset in fst format with an index over one of its columns
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
