% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{centroids}
\alias{centroids}
\title{centroids}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 3 rows and 5 columns.
}
\usage{
centroids
}
\description{
List of centroids
}
\keyword{datasets}
