% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSOA}
\alias{MSOA}
\title{MSOA}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 7264 rows and 2 columns.
}
\usage{
MSOA
}
\description{
Geographic Digital Boundaries in \code{sf} format of Medium Super Output Areas (MSOA) in England and Wales as of last Census 2021 (last update: Dec 2021).
}
\details{
The CRS is WGS84, EPSG:4326

For more details, consult the \href{https://geoportal.statistics.gov.uk/}{ONS Geoportal}
}
\keyword{datasets}
